+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project attempts to conform to [SemVer](https://semver.org/) guidelines whe
 ## [v0.2.0]
 * [Fixing wallet service interface](https://github.com/Ubunfu/mc-bounty-processor/pull/3)
 * [Adding bounty data load files](https://github.com/Ubunfu/mc-bounty-processor/pull/4)
+* [Adding a Discord notification when a bounty is turned in](https://github.com/Ubunfu/mc-bounty-processor/pull/5)
 
 ## [v0.1.0]
 * Initial implementation